move styled components in different files, create folders for each component
