+++ components/Form/styles.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const FormContainer = styled.form`
+  display: grid;
+  gap: 0.3rem;
+`;
