+++ components/Layout/styles.js
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+const Main = styled.main`
+  padding: 0.5rem;
+`;
