+++ components/MarkAsOwnedButton/styles.js
@@ -1,4 +1,3 @@
-import { CheckCircle, Circle } from "lucide-react";
 import styled from "styled-components";
 
 const Button = styled.button`
@@ -29,21 +28,3 @@ const OwnershipLabel = styled.span`
   font-weight: 700;
   letter-spacing: 0.4px;
 `;
-
-export default function MarkAsOwnedButton({ isOwned, onClick }) {
-  return (
-    <Button onClick={onClick} aria-label="Mark as owned" $isOwned={isOwned}>
-      {isOwned ? (
-        <>
-          <CheckCircle />
-          <OwnershipLabel>Owned</OwnershipLabel>
-        </>
-      ) : (
-        <>
-          <Circle />
-          <OwnershipLabel>Not Owned</OwnershipLabel>
-        </>
-      )}
-    </Button>
-  );
-}
